Data.Serialize

Description

The Serialize class

class Serialize t where

If your compiler has support for the DeriveGeneric and DefaultSignatures language extensions (ghc >= 7.2.1), the put and get methods will have default generic implementations.

To use this option, simply add a deriving Generic clause to your datatype and declare a Serialize instance for it without giving a definition for put and get.

Methods

put :: Putter t

Encode a value in the Put monad.

get :: Get t

Decode a value in the Get monad

Serialize serialisation

encode :: Serialize a => a -> ByteString

Encode a value using binary serialization to a strict ByteString.

encodeLazy :: Serialize a => a -> ByteString

Encode a value using binary serialization to a lazy ByteString.

decode :: Serialize a => ByteString -> Either String a

Decode a value from a strict ByteString, reconstructing the original structure.

decodeLazy :: Serialize a => ByteString -> Either String a

Decode a value from a lazy ByteString, reconstructing the original structure.

module Data.Serialize.Get

module Data.Serialize.Put

module Data.Serialize.IEEE754